# docs/swagger.json
{
    "schemes": [
        "https",
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Document API",
        "title": "Geolocation",
        "contact": {
            "name": "API Support",
            "url": "https://simpplify.com.br/contact",
            "email": "support@swagger.io"
        },
        "version": "1.0"
    },
    "paths": {
        "/address/city/{idState}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retorna todas as cidades de um estado específico, utilizando o ID do estado.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Endereços"
                ],
                "summary": "Buscar Cidades por Estado",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID do Estado",
                        "name": "idState",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de Cidades",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/address.CityResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Requisição Inválida",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Erro Interno do Servidor",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/address/find": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Busca um endereço por pesquisa. Pode ser: 1. Nome de localidade, 2. Latitude e Longitude ou 3. Endereço completo (Rua, bairro, número).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Endereços"
                ],
                "summary": "Buscar Endereço por Consulta",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Consulta do Endereço",
                        "name": "q",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Informações do Endereço",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/address.AddressResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Requisição Inválida",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Erro Interno do Servidor",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/address/find/v2": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Busca um endereço por pesquisa. Endereço completo (Rua, bairro, número).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Endereços"
                ],
                "summary": "Buscar Endereço por Consulta",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Consulta do Endereço",
                        "name": "q",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Informações do Endereço",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/address.AddressResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Requisição Inválida",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Erro Interno do Servidor",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/address/find/{cep}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Encontra um endereço pelo CEP, retornando o tipo baseado nas repetições encontradas.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Endereços"
                ],
                "summary": "Buscar Endereço por CEP",
                "parameters": [
                    {
                        "type": "string",
                        "description": "CEP",
                        "name": "cep",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Informações do Endereço",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/address.AddressCEPResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Requisição Inválida",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Erro Interno do Servidor",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/address/state": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retorna todos os estados disponíveis.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Endereços"
                ],
                "summary": "Buscar Todos os Estados",
                "responses": {
                    "200": {
                        "description": "Lista de Estados",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/address.StateResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Erro Interno do Servidor",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "address.AddressCEPResponse": {
            "type": "object",
            "properties": {
                "cep": {
                    "type": "string"
                },
                "city_name": {
                    "type": "string"
                },
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                },
                "neighborhood_name": {
                    "type": "string"
                },
                "state_uf": {
                    "type": "string"
                },
                "street_name": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "address.AddressDetail": {
            "type": "object",
            "properties": {
                "cep": {
                    "type": "string"
                },
                "id_address": {
                    "type": "integer"
                },
                "is_exactly": {
                    "type": "boolean"
                },
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                },
                "number": {
                    "type": "string"
                }
            }
        },
        "address.AddressResponse": {
            "type": "object",
            "properties": {
                "addresses": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/address.AddressDetail"
                    }
                },
                "city": {
                    "type": "string"
                },
                "id_street": {
                    "type": "integer"
                },
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                },
                "neighborhood": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                },
                "street": {
                    "type": "string"
                }
            }
        },
        "address.CityResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "address.StateResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "uf": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}